feat(auth): pluggable auth backend trait for SaaS repo support
Extract authentication into a pluggable AuthBackend trait so external
repos can provide custom auth providers (PropelAuth, Auth0, etc.)
without modifying OSS code. Also extract server entrypoint into a
reusable run() function and add external_id support to users/orgs.

Changes:
- Add AuthBackend trait (auth/backend.rs) and BuiltinAuthBackend impl
- Refactor AuthState to hold Arc<dyn AuthBackend>
- Extract server bootstrap into server::run() with factory pattern
- Add external_id column to users and organizations (migration 004)
- Add storage methods: get_user/org_by_external_id, upsert_org, ensure_membership
- Export AuthContext/OrgContext from UI providers for SaaS wrapper
- Add exports field to UI package.json
- Move seed and grpc_service to lib crate for reuse
