+++ apps/ui/package.json
@@ -2,6 +2,13 @@
   "name": "everruns-ui",
   "version": "0.6.0",
   "private": true,
+  "exports": {
+    "./providers/*": "./src/providers/*.tsx",
+    "./components/*": "./src/components/*.tsx",
+    "./hooks/*": "./src/hooks/*.ts",
+    "./lib/*": "./src/lib/*.ts",
+    "./lib/api/*": "./src/lib/api/*.ts"
+  },
   "scripts": {
     "dev": "next dev --port 9100",
     "build": "next build",
